Fix format specifier for ip_source bitfield in header_rewrite
Use PRIu64 with explicit cast to uint64_t to match the bitfield's underlying
type and ensure cross-platform compatibility. This fixes build failures on
Ubuntu 20.04.

+++ plugins/header_rewrite/conditions.cc
@@ -25,6 +25,7 @@
 #include <unistd.h>
 #include <arpa/inet.h>
 #include <cctype>
+#include <cinttypes>
 #include <sstream>
 #include <array>
 #include <atomic>
@@ -1768,7 +1769,7 @@ getClientAddr(TSHttpTxn txnp, int txn_private_slot)
     TSHttpTxnVerifiedAddrGet(txnp, &addr);
     break;
   default:
-    Dbg(pi_dbg_ctl, "Unknown IP source (%d) was specified", private_data.ip_source);
+    Dbg(pi_dbg_ctl, "Unknown IP source (%" PRIu64 ") was specified", static_cast<uint64_t>(private_data.ip_source));
     addr = TSHttpTxnClientAddrGet(txnp);
     break;
   }
